+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 * Add ``samples`` and ``shape`` attributes to :class:`pymaf.AnnFrame` class.
 * Rename :meth:`pymaf.MafFrame.compute_genes/tmb/waterfall` methods to :meth:`pymaf.MafFrame.matrix_genes/tmb/waterfall`.
 * Add ``keep_empty`` argument to :meth:`pymaf.MafFrame.matrix_waterfall/plot_oncoplot/plot_waterfall` methods.
+* Add :meth:`pymaf.MafFrame.filter_annot` method.
 
 0.14.0 (2021-06-20)
 -------------------